# internal/platform/docs/swagger.yaml
basePath: /api
definitions:
  httptransport.APIResponse:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  v1.AdminConfig:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
...
  v1.DatabaseConfig:
    properties:
      database:
        type: string
      host:
        type: string
      options:
        type: string
      password:
        type: string
      port:
        type: integer
      type:
        enum:
        - sqlite
        - mysql
        - postgresql
        type: string
      username:
        type: string
    required:
    - database
    - type
    type: object
...
  v1.DatabaseTestRequest:
    properties:
      config:
        $ref: '#/definitions/v1.DatabaseConfig'
    required:
    - config
    type: object
  v1.DatabaseTestResponse:
    properties:
      connected:
        type: boolean
      details: {}
      latency:
        description: 连接延迟（毫秒）
        type: integer
      message:
        type: string
      success:
        type: boolean
      version:
        type: string
    type: object
...
  v1.InitRequest:
    properties:
      admin_config:
        $ref: '#/definitions/v1.AdminConfig'
      database_config:
        $ref: '#/definitions/v1.DatabaseConfig'
      providers:
        additionalProperties: true
        type: object
      system_config: {}
    required:
    - admin_config
    - database_config
    type: object
...
  v1.Provider:
    properties:
      config:
        additionalProperties: true
        type: object
      created_at:
        type: string
      enabled:
        type: boolean
      metadata:
        additionalProperties: true
        type: object
      name:
        type: string
      status:
        description: active, inactive, error
        type: string
      type:
        description: llm, tts, asr, vllm
        type: string
      updated_at:
        type: string
    type: object
  v1.ProviderConfigRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      enabled:
        type: boolean
      name:
        type: string
      type:
        type: string
    required:
    - config
    - name
    - type
    type: object
  v1.ProviderList:
    properties:
      providers:
        items:
          $ref: '#/definitions/v1.Provider'
        type: array
    type: object
...
  v1.SystemConfigRequest:
    properties:
      setting:
        type: string
      value: {}
    type: object
  v1.SystemConfigResponse:
    properties:
      key:
        type: string
      type:
        description: string, number, boolean, object
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
      value: {}
    type: object
...
  webapi.AdminConfig:
    properties:
      email:
        type: string
      password:
        type: string
      type:
        description: '"random" 或 "custom"'
        type: string
      username:
        type: string
    type: object
...
  webapi.DatabaseConfig:
    properties:
      database:
        type: string
      host:
        type: string
      password:
        type: string
      path:
        description: SQLite数据库文件路径
        type: string
      port:
        type: integer
      type:
        type: string
      username:
        type: string
    type: object
...
  webapi.InitRequest:
    properties:
      adminConfig:
        $ref: '#/definitions/webapi.AdminConfig'
      databaseConfig:
        $ref: '#/definitions/webapi.DatabaseConfig'
      providers:
        additionalProperties: true
        type: object
      systemConfig: {}
    type: object
...
  webapi.SystemConfig:
    properties:
      prompt:
        type: string
      selectedASR:
        type: string
      selectedLLM:
        type: string
      selectedTTS:
        type: string
      selectedVLLLM:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: 小智服务端，包含OTA与Vision等接口
  title: 小智服务端 API 文档
  version: "1.0"
paths:
...
  /admin/system:
    get:
      description: 获取服务器的系统配置信息，包括选择的提供商和默认提示词
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/webapi.SystemConfig'
        "401":
          description: Unauthorized
          schema:
            type: object
      security:
      - BearerAuth: []
      summary: 获取系统配置
      tags:
      - Admin
...
  /admin/system/init:
    post:
      consumes:
      - application/json
      description: 初始化Xiaozhi-Flow系统，创建数据库和管理员用户
      parameters:
      - description: 初始化配置
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/webapi.InitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: object
      summary: 系统初始化
      tags:
      - Admin
...
  /providers:
    get:
      description: 获取所有配置的AI服务供应商
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.ProviderList'
              type: object
      summary: 获取供应商列表
      tags:
      - Providers
    post:
      consumes:
      - application/json
      description: 创建新的AI服务供应商配置
      parameters:
      - description: 供应商配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ProviderConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.Provider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 创建供应商配置
      tags:
      - Providers
  /providers/{type}/{name}:
    delete:
      description: 删除指定的供应商配置
      parameters:
      - description: 供应商类型
        in: path
        name: type
        required: true
        type: string
      - description: 供应商名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 删除供应商配置
      tags:
      - Providers
    put:
      consumes:
      - application/json
      description: 更新指定供应商的配置信息
      parameters:
      - description: 供应商类型
        in: path
        name: type
        required: true
        type: string
      - description: 供应商名称
        in: path
        name: name
        required: true
        type: string
      - description: 供应商配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.ProviderConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.Provider'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 更新供应商配置
      tags:
      - Providers
...
  /system/config:
    get:
      description: 获取系统配置项列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.SystemConfigResponse'
                  type: array
              type: object
      summary: 获取系统配置
      tags:
      - System
    post:
      consumes:
      - application/json
      description: 更新指定的系统配置项
      parameters:
      - description: 配置更新信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.SystemConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemConfigResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 更新系统配置
      tags:
      - System
...
  /system/config/test/database:
    post:
      consumes:
      - application/json
      description: 测试数据库连接配置
      parameters:
      - description: 数据库配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.DatabaseTestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/httptransport.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/v1.DatabaseTestResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 测试数据库连接
      tags:
      - System
...
  /system/init:
    get:
      description: 检查系统是否已完成初始化
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 获取系统初始化状态
      tags:
      - System
    post:
      consumes:
      - application/json
      description: 执行系统初始化操作
      parameters:
      - description: 初始化配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.InitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httptransport.APIResponse'
      summary: 系统初始化
      tags:
      - System
...
swagger: "2.0"