# internal/transport/http/v1/system_service.go
package v1

import (
	"fmt"
	"time"

	"github.com/gin-gonic/gin"
	"xiaozhi-server-go/internal/platform/config"
	"xiaozhi-server-go/internal/transport/http/types/v1"
	httpUtils "xiaozhi-server-go/internal/transport/http/utils"
	"xiaozhi-server-go/internal/utils"
)

// SystemServiceV1 V1版本系统服务
type SystemServiceV1 struct {
	logger *utils.Logger
	config *config.Config
	// TODO: 添加实际的业务逻辑依赖
}

// NewSystemServiceV1 创建系统服务V1实例
func NewSystemServiceV1(config *config.Config, logger *utils.Logger) (*SystemServiceV1, error) {
	if config == nil {
		return nil, fmt.Errorf("config is required")
	}
	if logger == nil {
		return nil, fmt.Errorf("logger is required")
	}

	return &SystemServiceV1{
		logger: logger,
		config: config,
	}, nil
}

// Register 注册系统API路由
func (s *SystemServiceV1) Register(router *gin.RouterGroup) {
	// 系统状态和健康检查
	system := router.Group("/system")
	{
		system.GET("/status", s.getSystemStatus)        // 获取系统状态
		system.GET("/health", s.healthCheck)            // 健康检查
		system.POST("/health", s.detailedHealthCheck)   // 详细健康检查
		system.GET("/time", s.getServerTime)            // 获取服务器时间
	}

	// 系统配置管理（需要管理员权限）
	configs := router.Group("/config")
	{
		configs.GET("", s.listConfigs)                 // 获取系统配置
		configs.POST("/test/database", s.testDatabase) // 测试数据库连接
		configs.GET("/schema/database", s.getDatabaseSchema) // 获取数据库模式
		configs.POST("", s.updateConfig)               // 更新系统配置
	}

	// 供应商管理
	providers := router.Group("/providers")
	{
		providers.GET("", s.listProviders)             // 获取供应商列表
		providers.POST("", s.createProvider)           // 创建供应商配置
		providers.PUT("/:type/:name", s.updateProvider) // 更新供应商配置
		providers.DELETE("/:type/:name", s.deleteProvider) // 删除供应商配置
	}

	// 系统初始化
	init := router.Group("/init")
	{
		init.GET("", s.getInitStatus)                 // 获取初始化状态
		init.POST("", s.initializeSystem)             // 系统初始化
	}
}

// getSystemStatus 获取系统状态
// @Summary 获取系统状态
// @Description 获取系统的运行状态和统计信息
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=v1.SystemStatus}
// @Router /system/status [get]
func (s *SystemServiceV1) getSystemStatus(c *gin.Context) {
	s.logger.InfoTag("API", "获取系统状态",
		"request_id", getRequestID(c),
	)

	// 模拟系统状态数据
	status := s.getMockSystemStatus()

	httpUtils.Response.Success(c, status, "获取系统状态成功")
}

// healthCheck 健康检查
// @Summary 健康检查
// @Description 快速健康检查
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=v1.HealthCheckResponse}
// @Router /system/health [get]
func (s *SystemServiceV1) healthCheck(c *gin.Context) {
	s.logger.InfoTag("API", "健康检查",
		"request_id", getRequestID(c),
	)

	// 模拟健康检查结果
	response := v1.HealthCheckResponse{
		Status:    "healthy",
		Timestamp: time.Now(),
		Duration:  15, // 15ms
		Checks: []v1.HealthCheckResult{
			{
				Name:     "database",
				Status:   "healthy",
				Duration: 5,
			},
			{
				Name:     "memory",
				Status:   "healthy",
				Duration: 3,
			},
			{
				Name:     "disk",
				Status:   "healthy",
				Duration: 7,
			},
		},
		Overall: "healthy",
	}

	httpUtils.Response.Success(c, response, "健康检查完成")
}

// detailedHealthCheck 详细健康检查
// @Summary 详细健康检查
// @Description 执行详细的健康检查，支持指定检查项目
// @Tags System
// @Accept json
// @Produce json
// @Param request body v1.HealthCheckRequest false "健康检查参数"
// @Success 200 {object} httptransport.APIResponse{data=v1.HealthCheckResponse}
// @Router /system/health [post]
func (s *SystemServiceV1) detailedHealthCheck(c *gin.Context) {
	var request v1.HealthCheckRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "详细健康检查",
		"checks", request.Checks,
		"timeout", request.Timeout,
		"request_id", getRequestID(c),
	)

	// 模拟详细健康检查
	checks := []v1.HealthCheckResult{
		{
			Name:     "database",
			Status:   "healthy",
			Message:  "数据库连接正常",
			Duration: 12,
			Details:  gin.H{"connections": 5, "queries_per_second": 120},
		},
		{
			Name:     "redis",
			Status:   "healthy",
			Message:  "Redis连接正常",
			Duration: 3,
			Details:  gin.H{"memory_usage": "45MB", "connected_clients": 3},
		},
	}

	if len(request.Checks) > 0 {
		// 过滤指定的检查项
		var filteredChecks []v1.HealthCheckResult
		for _, checkName := range request.Checks {
			for _, check := range checks {
				if check.Name == checkName {
					filteredChecks = append(filteredChecks, check)
					break
				}
			}
		}
		checks = filteredChecks
	}

	response := v1.HealthCheckResponse{
		Status:    "healthy",
		Timestamp: time.Now(),
		Duration:  25,
		Checks:    checks,
		Overall:   "healthy",
	}

	httpUtils.Response.Success(c, response, "详细健康检查完成")
}

// getServerTime 获取服务器时间
// @Summary 获取服务器时间
// @Description 获取服务器当前时间和相关信息
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=v1.ServerTimeInfo}
// @Router /system/time [get]
func (s *SystemServiceV1) getServerTime(c *gin.Context) {
	s.logger.InfoTag("API", "获取服务器时间",
		"request_id", getRequestID(c),
	)

	uptime := time.Since(time.Now().Add(-24 * time.Hour)).Seconds() // 模拟24小时运行时间
	serverTime := v1.ServerTimeInfo{
		CurrentTime: time.Now(),
		Timezone:    "UTC",
		Uptime:      int64(uptime),
		Load:        0.45, // 模拟CPU负载
	}

	httpUtils.Response.Success(c, serverTime, "获取服务器时间成功")
}

// listConfigs 获取系统配置
// @Summary 获取系统配置
// @Description 获取系统配置项列表
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=[]v1.SystemConfigResponse}
// @Router /system/config [get]
func (s *SystemServiceV1) listConfigs(c *gin.Context) {
	s.logger.InfoTag("API", "获取系统配置",
		"request_id", getRequestID(c),
	)

	// 模拟系统配置数据
	configs := []v1.SystemConfigResponse{
		{
			Key:       "app_name",
			Value:     "XiaoZhi Server",
			Type:      "string",
			UpdatedBy: "admin",
			UpdatedAt: time.Now().Add(-1 * time.Hour),
		},
		{
			Key:       "max_connections",
			Value:     1000,
			Type:      "number",
			UpdatedBy: "admin",
			UpdatedAt: time.Now().Add(-2 * time.Hour),
		},
		{
			Key:       "enable_metrics",
			Value:     true,
			Type:      "boolean",
			UpdatedBy: "admin",
			UpdatedAt: time.Now().Add(-30 * time.Minute),
		},
	}

	httpUtils.Response.Success(c, configs, "获取系统配置成功")
}

// testDatabase 测试数据库连接
// @Summary 测试数据库连接
// @Description 测试数据库连接配置
// @Tags System
// @Accept json
// @Produce json
// @Param request body v1.DatabaseTestRequest true "数据库配置"
// @Success 200 {object} httptransport.APIResponse{data=v1.DatabaseTestResponse}
// @Failure 400 {object} httptransport.APIResponse
// @Router /system/config/test/database [post]
func (s *SystemServiceV1) testDatabase(c *gin.Context) {
	var request v1.DatabaseTestRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "测试数据库连接",
		"type", request.Config.Type,
		"host", request.Config.Host,
		"database", request.Config.Database,
		"request_id", getRequestID(c),
	)

	// 模拟数据库连接测试
	response := v1.DatabaseTestResponse{
		Success:   true,
		Connected: true,
		Message:   "数据库连接成功",
		Latency:   25, // 25ms
		Version:   "8.0.28",
		Details: gin.H{
			"max_connections":     1000,
			"current_connections": 5,
			"charset":           "utf8mb4",
		},
	}

	httpUtils.Response.Success(c, response, "数据库连接测试完成")
}

// getDatabaseSchema 获取数据库模式
// @Summary 获取数据库模式
// @Description 获取数据库表结构信息
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=v1.DatabaseSchema}
// @Router /system/config/schema/database [get]
func (s *SystemServiceV1) getDatabaseSchema(c *gin.Context) {
	s.logger.InfoTag("API", "获取数据库模式",
		"request_id", getRequestID(c),
	)

	// 模拟数据库模式数据
	schema := v1.DatabaseSchema{
		Name: "xiaozhi_server",
		Tables: []v1.TableInfo{
			{
				Name:         "users",
				Engine:       "InnoDB",
				Charset:      "utf8mb4",
				Collation:    "utf8mb4_unicode_ci",
				Rows:         150,
				DataLength:   16384,
				IndexLength:  32768,
				AutoIncrement: 151,
				CreatedAt:    time.Now().Add(-30 * 24 * time.Hour),
				UpdatedAt:    time.Now().Add(-1 * time.Hour),
				Columns: []v1.ColumnInfo{
					{
						Name:         "id",
						Type:         "bigint",
						Nullable:     false,
						PrimaryKey:   true,
						AutoIncrement: true,
					},
					{
						Name:     "username",
						Type:     "varchar(50)",
						Nullable: false,
					},
					{
						Name:     "email",
						Type:     "varchar(100)",
						Nullable: false,
					},
					{
						Name:     "password_hash",
						Type:     "varchar(255)",
						Nullable: false,
					},
					{
						Name:         "created_at",
						Type:         "timestamp",
						Nullable:     false,
						Default:      "CURRENT_TIMESTAMP",
					},
				},
			},
		},
		Indexes: []v1.IndexInfo{
			{
				Name:      "PRIMARY",
				Type:      "btree",
				Unique:    true,
				Columns:   []string{"id"},
				TableName: "users",
			},
			{
				Name:      "idx_username",
				Type:      "btree",
				Unique:    true,
				Columns:   []string{"username"},
				TableName: "users",
			},
			{
				Name:      "idx_email",
				Type:      "btree",
				Unique:    true,
				Columns:   []string{"email"},
				TableName: "users",
			},
		},
		ForeignKeys: []v1.ForeignKeyInfo{},
	}

	httpUtils.Response.Success(c, schema, "获取数据库模式成功")
}

// updateConfig 更新系统配置
// @Summary 更新系统配置
// @Description 更新指定的系统配置项
// @Tags System
// @Accept json
// @Produce json
// @Param request body v1.SystemConfigRequest true "配置更新信息"
// @Success 200 {object} httptransport.APIResponse{data=v1.SystemConfigResponse}
// @Failure 400 {object} httptransport.APIResponse
// @Router /system/config [post]
func (s *SystemServiceV1) updateConfig(c *gin.Context) {
	var request v1.SystemConfigRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "更新系统配置",
		"setting", request.Setting,
		"request_id", getRequestID(c),
	)

	// 验证配置项是否存在
	validSettings := []string{"app_name", "max_connections", "enable_metrics", "debug_mode"}
	valid := false
	for _, setting := range validSettings {
		if request.Setting == setting {
			valid = true
			break
		}
	}

	if !valid {
		httpUtils.Response.Error(c, httpUtils.ErrorCodeInvalidInput, "无效的配置项")
		return
	}

	// 模拟配置更新
	response := v1.SystemConfigResponse{
		Key:       request.Setting,
		Value:     request.Value,
		Type:      "unknown", // 根据实际值类型判断
		UpdatedBy: "admin",   // 实际应该从认证信息中获取
		UpdatedAt: time.Now(),
	}

	httpUtils.Response.Success(c, response, "配置更新成功")
}

// listProviders 获取供应商列表
// @Summary 获取供应商列表
// @Description 获取所有配置的AI服务供应商
// @Tags Providers
// @Produce json
// @Success 200 {object} httptransport.APIResponse{data=v1.ProviderList}
// @Router /providers [get]
func (s *SystemServiceV1) listProviders(c *gin.Context) {
	s.logger.InfoTag("API", "获取供应商列表",
		"request_id", getRequestID(c),
	)

	// 模拟供应商数据
	providers := []v1.Provider{
		{
			Type:     "llm",
			Name:     "openai",
			Status:   "active",
			Config:   gin.H{"api_key": "sk-***", "model": "gpt-3.5-turbo"},
			Metadata: gin.H{"description": "OpenAI GPT模型"},
			Enabled:  true,
			CreatedAt: time.Now().Add(-7 * 24 * time.Hour),
			UpdatedAt: time.Now().Add(-1 * time.Hour),
		},
		{
			Type:     "tts",
			Name:     "azure-tts",
			Status:   "active",
			Config:   gin.H{"api_key": "***", "region": "eastus"},
			Metadata: gin.H{"description": "Azure文字转语音"},
			Enabled:  true,
			CreatedAt: time.Now().Add(-5 * 24 * time.Hour),
			UpdatedAt: time.Now().Add(-2 * time.Hour),
		},
		{
			Type:     "asr",
			Name:     "whisper",
			Status:   "inactive",
			Config:   gin.H{},
			Metadata: gin.H{"description": "OpenAI Whisper语音识别"},
			Enabled:  false,
			CreatedAt: time.Now().Add(-3 * 24 * time.Hour),
			UpdatedAt: time.Now().Add(-1 * time.Hour),
		},
	}

	response := v1.ProviderList{
		Providers: providers,
	}

	httpUtils.Response.Success(c, response, "获取供应商列表成功")
}

// createProvider 创建供应商配置
// @Summary 创建供应商配置
// @Description 创建新的AI服务供应商配置
// @Tags Providers
// @Accept json
// @Produce json
// @Param request body v1.ProviderConfigRequest true "供应商配置"
// @Success 201 {object} httptransport.APIResponse{data=v1.Provider}
// @Failure 400 {object} httptransport.APIResponse
// @Router /providers [post]
func (s *SystemServiceV1) createProvider(c *gin.Context) {
	var request v1.ProviderConfigRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "创建供应商配置",
		"type", request.Type,
		"name", request.Name,
		"enabled", request.Enabled,
		"request_id", getRequestID(c),
	)

	// 检查供应商是否已存在
	if s.providerExists(request.Type, request.Name) {
		httpUtils.Response.Error(c, httpUtils.ErrorCodeProviderExists, "供应商已存在")
		return
	}

	// 模拟创建供应商
	provider := v1.Provider{
		Type:     request.Type,
		Name:     request.Name,
		Status:   "active",
		Config:   request.Config,
		Metadata: gin.H{},
		Enabled:  request.Enabled,
		CreatedAt: time.Now(),
		UpdatedAt: time.Now(),
	}

	httpUtils.Response.Created(c, provider, "供应商创建成功")
}

// updateProvider 更新供应商配置
// @Summary 更新供应商配置
// @Description 更新指定供应商的配置信息
// @Tags Providers
// @Accept json
// @Produce json
// @Param type path string true "供应商类型"
// @Param name path string true "供应商名称"
// @Param request body v1.ProviderConfigRequest true "供应商配置"
// @Success 200 {object} httptransport.APIResponse{data=v1.Provider}
// @Failure 400 {object} httptransport.APIResponse
// @Failure 404 {object} httptransport.APIResponse
// @Router /providers/{type}/{name} [put]
func (s *SystemServiceV1) updateProvider(c *gin.Context) {
	providerType := c.Param("type")
	providerName := c.Param("name")

	if providerType == "" || providerName == "" {
		httpUtils.Response.BadRequest(c, "供应商类型和名称不能为空")
		return
	}

	var request v1.ProviderConfigRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "更新供应商配置",
		"type", providerType,
		"name", providerName,
		"enabled", request.Enabled,
		"request_id", getRequestID(c),
	)

	// 检查供应商是否存在
	if !s.providerExists(providerType, providerName) {
		httpUtils.Response.NotFound(c, "供应商")
		return
	}

	// 模拟更新供应商
	provider := v1.Provider{
		Type:     providerType,
		Name:     providerName,
		Status:   "active",
		Config:   request.Config,
		Metadata: gin.H{},
		Enabled:  request.Enabled,
		CreatedAt: time.Now().Add(-7 * 24 * time.Hour),
		UpdatedAt: time.Now(),
	}

	httpUtils.Response.Success(c, provider, "供应商配置更新成功")
}

// deleteProvider 删除供应商配置
// @Summary 删除供应商配置
// @Description 删除指定的供应商配置
// @Tags Providers
// @Produce json
// @Param type path string true "供应商类型"
// @Param name path string true "供应商名称"
// @Success 200 {object} httptransport.APIResponse
// @Failure 404 {object} httptransport.APIResponse
// @Router /providers/{type}/{name} [delete]
func (s *SystemServiceV1) deleteProvider(c *gin.Context) {
	providerType := c.Param("type")
	providerName := c.Param("name")

	if providerType == "" || providerName == "" {
		httpUtils.Response.BadRequest(c, "供应商类型和名称不能为空")
		return
	}

	s.logger.InfoTag("API", "删除供应商配置",
		"type", providerType,
		"name", providerName,
		"request_id", getRequestID(c),
	)

	// 检查供应商是否存在
	if !s.providerExists(providerType, providerName) {
		httpUtils.Response.NotFound(c, "供应商")
		return
	}

	httpUtils.Response.Success(c, gin.H{"provider": fmt.Sprintf("%s/%s", providerType, providerName)}, "供应商删除成功")
}

// getInitStatus 获取初始化状态
// @Summary 获取系统初始化状态
// @Description 检查系统是否已完成初始化
// @Tags System
// @Produce json
// @Success 200 {object} httptransport.APIResponse
// @Router /system/init [get]
func (s *SystemServiceV1) getInitStatus(c *gin.Context) {
	s.logger.InfoTag("API", "获取初始化状态",
		"request_id", getRequestID(c),
	)

	// 模拟初始化状态检查
	status := gin.H{
		"initialized": true,
		"version":     "1.0.0",
		"init_time":   time.Now().Add(-30 * 24 * time.Hour),
		"admin_created": true,
		"database_configured": true,
		"providers_configured": 2,
	}

	httpUtils.Response.Success(c, status, "获取初始化状态成功")
}

// initializeSystem 系统初始化
// @Summary 系统初始化
// @Description 执行系统初始化操作
// @Tags System
// @Accept json
// @Produce json
// @Param request body v1.InitRequest true "初始化配置"
// @Success 200 {object} httptransport.APIResponse
// @Failure 400 {object} httptransport.APIResponse
// @Router /system/init [post]
func (s *SystemServiceV1) initializeSystem(c *gin.Context) {
	var request v1.InitRequest
	if err := c.ShouldBindJSON(&request); err != nil {
		httpUtils.Response.ValidationError(c, err)
		return
	}

	s.logger.InfoTag("API", "系统初始化",
		"database_type", request.DatabaseConfig.Type,
		"admin_username", request.AdminConfig.Username,
		"request_id", getRequestID(c),
	)

	// 模拟系统初始化
	result := gin.H{
		"success": true,
		"message": "系统初始化完成",
		"database": gin.H{
			"connected": true,
			"tables_created": true,
		},
		"admin": gin.H{
			"created": true,
			"user_id": 1,
		},
		"providers_count": len(request.Providers),
	}

	httpUtils.Response.Success(c, result, "系统初始化成功")
}

// ========== 模拟数据方法 ==========
// TODO: 实际实现中应该从数据库或配置中获取真实数据

func (s *SystemServiceV1) getMockSystemStatus() v1.SystemStatus {
	return v1.SystemStatus{
		Status:      "running",
		Version:     "1.0.0",
		StartTime:   time.Now().Add(-24 * time.Hour),
		Uptime:      86400, // 24小时
		Environment: "production",
		GoVersion:   "1.21.0",
		Services: []v1.ServiceStatus{
			{
				Name:      "web-server",
				Status:    "running",
				Health:    "healthy",
				Uptime:    86400,
				StartTime: time.Now().Add(-24 * time.Hour),
			},
			{
				Name:      "database",
				Status:    "running",
				Health:    "healthy",
				Uptime:    86400,
				StartTime: time.Now().Add(-24 * time.Hour),
			},
			{
				Name:      "redis",
				Status:    "running",
				Health:    "healthy",
				Uptime:    86400,
				StartTime: time.Now().Add(-24 * time.Hour),
			},
		},
		Database: &v1.DatabaseStatus{
			Status:         "connected",
			Connection:     "connected",
			Type:           "mysql",
			Host:           "localhost",
			Port:           3306,
			Database:       "xiaozhi_server",
			MaxConnections: 1000,
			OpenConnections: 5,
			SlowQueries:    0,
		},
		Statistics: v1.SystemStatistics{
			TotalRequests:    100000,
			SuccessRequests:  95000,
			ErrorRequests:    5000,
			ActiveSessions:   120,
			TotalUsers:       150,
			RegisteredUsers:  120,
			StorageUsed:      2147483648, // 2GB
			MemoryUsage:      134217728,  // 128MB
			CPUUsage:         25.5,
			DiskUsage:        65.2,
		},
	}
}

func (s *SystemServiceV1) providerExists(providerType, providerName string) bool {
	// 简单模拟供应商存在性检查
	existingProviders := []string{
		"llm/openai",
		"tts/azure-tts",
		"asr/whisper",
	}

	providerKey := fmt.Sprintf("%s/%s", providerType, providerName)
	for _, existing := range existingProviders {
		if existing == providerKey {
			return true
		}
	}
	return false
}